更改用户等级、管理列表的逻辑！完善了类目，级联可以列出三级的信息。

+++ OfficeAutomation/BLL/BLL_Classtime_All.cs
@@ -118,8 +118,8 @@ namespace OfficeAutomation.BLL
                 return result;
             }
             result.code = 0;
-            result.count = dalClasstimeAll.List(users.collegeid).Count;
-            result.data = dalClasstimeAll.List(users.collegeid,page, limit);
+            result.count = dalClasstimeAll.List((int)users.collegeid).Count;
+            result.data = dalClasstimeAll.List((int)users.collegeid,page, limit);
             return result;
         }
         /// <summary>
